fix(memory-wiki): rekey legacy rows before JSON import and count hidden rows in capacity preflight
A full raw legacy namespace crashed the JSON-to-plugin-state migration
mid-write: store.read hides unscoped rows, so the capacity preflight
undercounted physical rows. Reorder the doctor migrations so ownership
rekeying runs first, account for hidden unscoped rows in the JSON
preflight, and split the legacy migration into its own module.
