+++ extensions/memory-wiki/src/source-sync-state.ts
@@ -11,11 +11,10 @@ import type {
 import { FsSafeError, root as fsRoot } from "openclaw/plugin-sdk/security-runtime";
 import { isRecord } from "openclaw/plugin-sdk/string-coerce-runtime";
 import { createWikiPageFilename, extractHumanNotesBlock } from "./markdown.js";
-import { resolveUnsafeLocalPagePath } from "./source-path-shared.js";
 
 export type MemoryWikiImportedSourceGroup = "bridge" | "unsafe-local";
 
-type MemoryWikiImportedSourceStateEntry = {
+export type MemoryWikiImportedSourceStateEntry = {
   group: MemoryWikiImportedSourceGroup;
   pagePath: string;
   sourcePath: string;
@@ -48,7 +47,7 @@ type MemoryWikiSourceSyncStateStore = {
   ) => Promise<void>;
 };
 
-type MemoryWikiSourceSyncStateRecord = MemoryWikiImportedSourceStateEntry & {
+export type MemoryWikiSourceSyncStateRecord = MemoryWikiImportedSourceStateEntry & {
   vaultRootKey: string;
   syncKey: string;
 };
@@ -70,13 +69,13 @@ export function scopeImportedSourceSyncKey(
   return `${group}:${bindingKey}`;
 }
 
-function isScopedImportedSourceSyncKey(syncKey: string): boolean {
+export function isScopedImportedSourceSyncKey(syncKey: string): boolean {
   return syncKey.startsWith("bridge:") || syncKey.startsWith("unsafe-local:");
 }
 
 // The plugin SDK does not expose PluginStateStoreError; match the documented
 // reject-new capacity contract structurally instead of widening SDK surface.
-function isPluginStateLimitExceeded(error: unknown): boolean {
+export function isPluginStateLimitExceeded(error: unknown): boolean {
   return isRecord(error) && error.code === "PLUGIN_STATE_LIMIT_EXCEEDED";
 }
 const MAX_MEMORY_WIKI_NOTES_RECOVERY_BYTES = 16 * 1024 * 1024;
@@ -108,7 +107,7 @@ function cloneSourceSyncState(state: MemoryWikiImportedSourceState): MemoryWikiI
   };
 }
 
-function normalizeSourceSyncState(value: unknown): MemoryWikiImportedSourceState {
+export function normalizeSourceSyncState(value: unknown): MemoryWikiImportedSourceState {
   if (!value || typeof value !== "object" || Array.isArray(value)) {
     return EMPTY_STATE;
   }
@@ -143,11 +142,11 @@ function normalizeSourceSyncState(value: unknown): MemoryWikiImportedSourceState
   return { version: 1, entries };
 }
 
-function resolveVaultRootKey(vaultRoot: string): string {
+export function resolveVaultRootKey(vaultRoot: string): string {
   return createHash("sha256").update(path.resolve(vaultRoot), "utf8").digest("hex").slice(0, 32);
 }
 
-function resolveStateEntryKey(vaultRootKey: string, syncKey: string): string {
+export function resolveStateEntryKey(vaultRootKey: string, syncKey: string): string {
   return createHash("sha256").update(`${vaultRootKey}\0${syncKey}`, "utf8").digest("hex");
 }
 
@@ -589,214 +588,4 @@ export function setImportedSourceEntry(params: {
   const changes = sourceSyncStateChanges.get(params.state);
   changes?.deleteKeys.delete(params.syncKey);
   changes?.upsertKeys.add(params.syncKey);
-}
-
-type LegacyImportedSourceSyncRow = {
-  storeKey: string;
-  syncKey: string;
-  entry: MemoryWikiImportedSourceStateEntry;
-};
-
-async function listLegacyImportedSourceSyncRows(params: {
-  vaultRoot: string;
-  openKeyedStore: <T>(options: OpenKeyedStoreOptions) => PluginStateKeyedStore<T>;
-}): Promise<{ legacyRows: LegacyImportedSourceSyncRow[]; retainedKeys: string[] }> {
-  const raw = params.openKeyedStore<MemoryWikiSourceSyncStateRecord>({
-    namespace: MEMORY_WIKI_SOURCE_SYNC_STATE_NAMESPACE,
-    maxEntries: MEMORY_WIKI_SOURCE_SYNC_STATE_MAX_ENTRIES,
-    overflowPolicy: "reject-new",
-  });
-  const vaultRootKey = resolveVaultRootKey(params.vaultRoot);
-  const legacyRows: LegacyImportedSourceSyncRow[] = [];
-  const retainedKeys: string[] = [];
-  for (const row of await raw.entries()) {
-    const value = row.value;
-    if (
-      value.vaultRootKey !== vaultRootKey ||
-      typeof value.syncKey !== "string" ||
-      isScopedImportedSourceSyncKey(value.syncKey)
-    ) {
-      continue;
-    }
-    const normalized = normalizeSourceSyncState({
-      version: 1,
-      entries: { [value.syncKey]: value },
-    });
-    const entry = normalized.entries[value.syncKey];
-    if (!entry) {
-      // Unknown shapes stay untouched: no runtime version can read them, and
-      // deleting them is a data call the operator should see in warnings first.
-      retainedKeys.push(value.syncKey);
-      continue;
-    }
-    legacyRows.push({ storeKey: row.key, syncKey: value.syncKey, entry });
-  }
-  return { legacyRows, retainedKeys };
-}
-
-export async function countLegacyImportedSourceSyncRows(params: {
-  vaultRoot: string;
-  openKeyedStore: <T>(options: OpenKeyedStoreOptions) => PluginStateKeyedStore<T>;
-}): Promise<number> {
-  const { legacyRows } = await listLegacyImportedSourceSyncRows(params);
-  return legacyRows.length;
-}
-
-type LegacyImportedSourceSyncMigrationResult = {
-  translatedCount: number;
-  prunedCount: number;
-  retainedKeys: string[];
-  // Legacy rows kept because a full reject-new namespace could not hold the
-  // scoped replacement; deleting them first would risk losing durable
-  // ownership between two separate write transactions.
-  capacityRetainedKeys: string[];
-};
-
-export function translateLegacyImportedSourceSyncKey(params: {
-  entry: MemoryWikiImportedSourceStateEntry;
-  syncKey: string;
-  unsafeLocalConfiguredRoots: readonly string[];
-}): string | undefined {
-  if (params.entry.group === "bridge") {
-    // Legacy rows carry no workspace binding, so they cannot be re-keyed to
-    // the page binding here. The translated row keeps the page owned until the
-    // next bridge sync re-owns it under the binding key; the shared-page prune
-    // guard then retires this row without deleting the page.
-    return scopeImportedSourceSyncKey("bridge", params.syncKey);
-  }
-  // The translated row must own the page the legacy row recorded: when nested
-  // roots both contain the source, only the root whose page identity matches
-  // the recorded page is the original binding. Anything else is a root move
-  // and must go through the stale salvage path instead.
-  const root = params.unsafeLocalConfiguredRoots.find(
-    (candidate) =>
-      isPathInside(candidate, params.entry.sourcePath) &&
-      resolveUnsafeLocalPagePath({
-        configuredPath: candidate,
-        absolutePath: params.entry.sourcePath,
-      }).pagePath === params.entry.pagePath,
-  );
-  return root
-    ? scopeImportedSourceSyncKey("unsafe-local", `${root}\0${params.entry.sourcePath}`)
-    : undefined;
-}
-
-export async function pruneLegacyImportedSourceRows(params: {
-  vaultRoot: string;
-  openKeyedStore: <T>(options: OpenKeyedStoreOptions) => PluginStateKeyedStore<T>;
-  rows: ReadonlyArray<{
-    syncKey: string;
-    entry: MemoryWikiImportedSourceStateEntry;
-  }>;
-}): Promise<number> {
-  if (params.rows.length === 0) {
-    return 0;
-  }
-  const store = createMemoryWikiSourceSyncStateStore(params.openKeyedStore);
-  const state = await readMemoryWikiSourceSyncState(params.vaultRoot, store);
-  for (const row of params.rows) {
-    state.entries[row.syncKey] = row.entry;
-  }
-  // Keep every live scoped row; only the injected stale legacy rows prune.
-  const activeKeys = new Set(
-    Object.keys(state.entries).filter((key) => isScopedImportedSourceSyncKey(key)),
-  );
-  let prunedCount = 0;
-  for (const group of new Set(params.rows.map((row) => row.entry.group))) {
-    prunedCount += await pruneImportedSourceEntries({
-      vaultRoot: params.vaultRoot,
-      group,
-      activeKeys,
-      state,
-    });
-  }
-  await writeMemoryWikiSourceSyncState(params.vaultRoot, state, store);
-  return prunedCount;
-}
-
-export async function migrateLegacyImportedSourceSyncKeys(params: {
-  vaultRoot: string;
-  openKeyedStore: <T>(options: OpenKeyedStoreOptions) => PluginStateKeyedStore<T>;
-  unsafeLocalConfiguredPaths: readonly string[];
-}): Promise<LegacyImportedSourceSyncMigrationResult> {
-  const raw = params.openKeyedStore<MemoryWikiSourceSyncStateRecord>({
-    namespace: MEMORY_WIKI_SOURCE_SYNC_STATE_NAMESPACE,
-    maxEntries: MEMORY_WIKI_SOURCE_SYNC_STATE_MAX_ENTRIES,
-    overflowPolicy: "reject-new",
-  });
-  const vaultRootKey = resolveVaultRootKey(params.vaultRoot);
-  const { legacyRows, retainedKeys } = await listLegacyImportedSourceSyncRows(params);
-  const configuredRoots = params.unsafeLocalConfiguredPaths.map((configuredPath) =>
-    path.resolve(configuredPath),
-  );
-
-  // Migrates one legacy row to its scoped key. Returns "capacity" when a full
-  // reject-new namespace cannot hold the temporary extra row; the legacy row
-  // is left untouched in that case so durable ownership is never lost.
-  const migrateRow = async (
-    row: LegacyImportedSourceSyncRow,
-    nextSyncKey: string,
-  ): Promise<"migrated" | "capacity"> => {
-    const nextStoreKey = resolveStateEntryKey(vaultRootKey, nextSyncKey);
-    if (await raw.lookup(nextStoreKey)) {
-      await raw.delete(row.storeKey);
-      return "migrated";
-    }
-    try {
-      // Register the scoped replacement before deleting the legacy row so a
-      // crash mid-migration never loses ownership; reruns are idempotent.
-      await raw.register(nextStoreKey, { ...row.entry, vaultRootKey, syncKey: nextSyncKey });
-    } catch (error) {
-      if (!isPluginStateLimitExceeded(error)) {
-        throw error;
-      }
-      return "capacity";
-    }
-    await raw.delete(row.storeKey);
-    return "migrated";
-  };
-
-  let translatedCount = 0;
-  const staleRows: LegacyImportedSourceSyncRow[] = [];
-  const capacityRows: { row: LegacyImportedSourceSyncRow; nextSyncKey: string }[] = [];
-  const capacityRetainedKeys: string[] = [];
-  for (const row of legacyRows) {
-    const nextSyncKey = translateLegacyImportedSourceSyncKey({
-      entry: row.entry,
-      syncKey: row.syncKey,
-      unsafeLocalConfiguredRoots: configuredRoots,
-    });
-    if (!nextSyncKey) {
-      // The recorded binding no longer matches any configured unsafe-local
-      // root: prune through the canonical salvage path, never drop silently.
-      staleRows.push(row);
-      continue;
-    }
-    if ((await migrateRow(row, nextSyncKey)) === "migrated") {
-      translatedCount += 1;
-    } else {
-      capacityRows.push({ row, nextSyncKey });
-    }
-  }
-
-  // Prune stale rows before retrying capacity-blocked ones: pruning frees
-  // namespace slots that may let those translations complete in the same run.
-  const prunedCount = await pruneLegacyImportedSourceRows({
-    vaultRoot: params.vaultRoot,
-    openKeyedStore: params.openKeyedStore,
-    rows: staleRows,
-  });
-
-  for (const { row, nextSyncKey } of capacityRows) {
-    if ((await migrateRow(row, nextSyncKey)) === "migrated") {
-      translatedCount += 1;
-      continue;
-    }
-    // Still no free slot. Deleting the legacy row first would leave a crash
-    // window with no durable ownership at all, so retain it instead; Doctor
-    // surfaces a warning and a later run retries once capacity frees up.
-    capacityRetainedKeys.push(row.syncKey);
-  }
-
-  return { translatedCount, prunedCount, retainedKeys, capacityRetainedKeys };
 }
